dockerized-stock-pipeline-dagster/dagster_project/requirements.txt
dagster==1.7.11
dagster-webserver==1.7.11
dagster-postgres==0.23.11
requests==2.32.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
